Change int to size_t

+++ src/shaders.cpp
@@ -29,8 +29,8 @@ GLuint compile_shader(std::string source, GLenum shaderType)
 // Compile shaders and link program
 GLuint build_program()
 {
-	std::string source_vs(source_vs_start, (int)source_vs_size);
-	std::string source_fs(source_fs_start, (int)source_fs_size);
+	std::string source_vs(source_vs_start, (size_t)source_vs_size);
+	std::string source_fs(source_fs_start, (size_t)source_fs_size);
 
 	GLuint vs = compile_shader(source_vs, GL_VERTEX_SHADER);
 	GLuint fs = compile_shader(source_fs, GL_FRAGMENT_SHADER);
